pyproject.toml
[project]
name = "slide2note"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.1.0",
    "easyocr>=1.7.2",
    "google>=3.0.0",
    "google-genai>=1.16.1",
    "kaggle>=1.7.4.5",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "pyocr>=0.8.5",
    "pypdf>=5.5.0",
    "python-dotenv>=1.1.0",
    "torch>=2.7.0",
]
